Added UNIFORM BUFFER OBJECT For often reused uniform data (viewMat, ProjMat, lightData, etc). Allows for different shader programs to access same uniforms from gpu memory block. So modifying a single buffer is enough to update specified uniforms in multiple shaders. Otherwise we would need calls to bind uniforms and assign the same data for each shader individually.
NOTE: This uses a very specific memory layout called std140.
The compiler is not allowed to pack this layout for optimization,
to ensure it stays the same for all shader programs that use it.
MAKE SURE TO MAINTAIN UNIFORM BLOCK LAYOUT ACROSS ALL SHADERS FILES!
THIS LAYOUT ONLY ALLOWS VECTORS TO BE VEC2 OR VEC4 !!
(VEC3 ARE PADDED, BUT DONT RELY ON GL IMPLEMENTATION FOR IT)
ORDER OF UNIFORMS WITHIN BLOCK MUST BE CONSISTENT !